+++ tests/state_and_value_merge.rs
@@ -22,7 +22,7 @@ fn input_level() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<InputLevel>()?;
-    assert_eq!(action.value, (Vec2::Y * 2.0).into());
+    assert_eq!(action.value, Vec2::Y * 2.0);
     assert_eq!(action.state, ActionState::Ongoing);
 
     app.world_mut()
@@ -33,7 +33,7 @@ fn input_level() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<InputLevel>()?;
-    assert_eq!(action.value, (Vec2::Y * 2.0).into());
+    assert_eq!(action.value, Vec2::Y * 2.0);
     assert_eq!(action.state, ActionState::Fired);
 
     let mut keys = app.world_mut().resource_mut::<ButtonInput<KeyCode>>();
@@ -44,7 +44,7 @@ fn input_level() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<InputLevel>()?;
-    assert_eq!(action.value, Vec2::NEG_Y.into());
+    assert_eq!(action.value, Vec2::NEG_Y);
     assert_eq!(action.state, ActionState::Fired);
 
     app.world_mut()
@@ -55,7 +55,7 @@ fn input_level() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<InputLevel>()?;
-    assert_eq!(action.value, Vec2::Y.into());
+    assert_eq!(action.value, Vec2::Y);
     assert_eq!(action.state, ActionState::Fired);
 
     app.world_mut()
@@ -66,7 +66,7 @@ fn input_level() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<InputLevel>()?;
-    assert_eq!(action.value, Vec2::ZERO.into());
+    assert_eq!(action.value, Vec2::ZERO);
     assert_eq!(
         action.state,
         ActionState::None,
@@ -96,7 +96,7 @@ fn action_level() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<ActionLevel>()?;
-    assert_eq!(action.value, (Vec2::NEG_Y * 2.0).into());
+    assert_eq!(action.value, Vec2::NEG_Y * 2.0);
     assert_eq!(action.state, ActionState::Ongoing);
 
     app.world_mut()
@@ -107,7 +107,7 @@ fn action_level() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<ActionLevel>()?;
-    assert_eq!(action.value, (Vec2::NEG_Y * 2.0).into());
+    assert_eq!(action.value, Vec2::NEG_Y * 2.0);
     assert_eq!(action.state, ActionState::Fired);
 
     let mut keys = app.world_mut().resource_mut::<ButtonInput<KeyCode>>();
@@ -118,7 +118,7 @@ fn action_level() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<ActionLevel>()?;
-    assert_eq!(action.value, (Vec2::NEG_Y * 2.0).into());
+    assert_eq!(action.value, Vec2::NEG_Y * 2.0);
     assert_eq!(action.state, ActionState::Fired);
 
     app.world_mut()
@@ -129,7 +129,7 @@ fn action_level() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<ActionLevel>()?;
-    assert_eq!(action.value, (Vec2::NEG_Y * 4.0).into());
+    assert_eq!(action.value, Vec2::NEG_Y * 4.0);
     assert_eq!(action.state, ActionState::Fired);
 
     app.world_mut()
@@ -140,7 +140,7 @@ fn action_level() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<ActionLevel>()?;
-    assert_eq!(action.value, (Vec2::NEG_Y * 4.0).into());
+    assert_eq!(action.value, Vec2::NEG_Y * 4.0);
     assert_eq!(
         action.state,
         ActionState::None,
@@ -170,7 +170,7 @@ fn both_levels() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<BothLevels>()?;
-    assert_eq!(action.value, (Vec2::Y * 2.0).into());
+    assert_eq!(action.value, Vec2::Y * 2.0);
     assert_eq!(action.state, ActionState::Ongoing);
 
     app.world_mut()
@@ -181,7 +181,7 @@ fn both_levels() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<BothLevels>()?;
-    assert_eq!(action.value, (Vec2::Y * 2.0).into());
+    assert_eq!(action.value, Vec2::Y * 2.0);
     assert_eq!(action.state, ActionState::Fired);
 
     let mut keys = app.world_mut().resource_mut::<ButtonInput<KeyCode>>();
@@ -192,7 +192,7 @@ fn both_levels() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<BothLevels>()?;
-    assert_eq!(action.value, Vec2::NEG_Y.into());
+    assert_eq!(action.value, Vec2::NEG_Y);
     assert_eq!(action.state, ActionState::Fired);
 
     app.world_mut()
@@ -203,7 +203,7 @@ fn both_levels() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<BothLevels>()?;
-    assert_eq!(action.value, Vec2::Y.into());
+    assert_eq!(action.value, Vec2::Y);
     assert_eq!(action.state, ActionState::Fired);
 
     app.world_mut()
@@ -214,7 +214,7 @@ fn both_levels() -> Result<()> {
 
     let actions = app.world().get::<Actions<Test>>(entity).unwrap();
     let action = actions.get::<BothLevels>()?;
-    assert_eq!(action.value, Vec2::Y.into());
+    assert_eq!(action.value, Vec2::Y);
     assert_eq!(
         action.state,
         ActionState::None,